change background color to white, fixed "Can't see description if title of news doesn't fit in two lines" bug
