+++ app/src/main/res/layout/fragment_news.xml
@@ -4,6 +4,7 @@
     android:id="@+id/recyclerView"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/whiteBackground"
     android:overScrollMode="never"
     android:scrollbars="vertical"
     app:layoutManager="android.support.v7.widget.LinearLayoutManager" />
