+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="bottomNavigationTimerTab">#FFFF8800</color>
     <color name="bottomNavigationScheduleTab">#FF5252</color>
     <color name="background">#EEEEEE</color>
+    <color name="whiteBackground">#FFFFFF</color>
 </resources>
